Remove character "-" from vsphere_vm roles
at the end of main file, depends from which deployment method you choose dhcp or static, a second playbook must be called. 

Here this character should not be there, otherwise in scenario 5, at the end, it will throw the error "could not run dhcp-.yaml"

This character is already present in the line for calling static-* playbooks.